fix(checker): preserve class unique symbol keyof keys
