+++ crates/tsz-checker/Cargo.toml
@@ -622,6 +622,10 @@ name = "keyof_well_known_symbol_key_tests"
 path = "tests/keyof_well_known_symbol_key_tests.rs"
 
 [[test]]
+name = "keyof_class_unique_symbol_key_tests"
+path = "tests/keyof_class_unique_symbol_key_tests.rs"
+
+[[test]]
 name = "partial_self_argument_diagnostic_tests"
 path = "tests/partial_self_argument_diagnostic_tests.rs"
